GUI overhaul.
MainGUI is the central JFrame of the application.
The currentpanel JPanel on it holds the current selected tab.
Previous QueryQueries(now BrowseGUI), CreateQuestionGUI, UserProfileGUI
and UserManagementGUI are now of JPanel class and are stored on the
gui.panels package.

The gui.components package now holds the various custom Swing components
used in the project.