Browse window somewhat done.
THE DATABASE IS SET TO REMOTE PLEASE TEST IF IT WORKS FOR YOU. THE
CONFIGURATION IS IN THE resources/log/config.xml file IF IT DOESNT WORK
JUST SET local TO TRUE AND WORK LOCALLY.

Introduced multiple betting.
Updated Bet class, now bets can involve multiple predictions and they
belong to a type that can be an accumulator(Single, Double, Treble...)
or a full cover(Trixie, Patent..).

Updated user management tab, now the searches are performed
automatically when typing in the search bar, and a few matching
selections are available(full match,beginning with, contains word).

Updated create question tab, now answers are added to a table which
allows to view the selection of answers the question will have. Upon
pressing the create button the question will be created with those
answers. It's no longer possible to have more than one answer with the
same value (answers can have the same odds however).